Remove context off tchannel inbound call struct
