+++ codegen/templates/tchannel_endpoint.tmpl
@@ -44,7 +44,6 @@ func New{{$handlerName}}(deps *module.Dependencies) *{{$handlerName}} {
 		Deps: deps,
 	}
 	handler.endpoint = zanzibar.NewTChannelEndpoint(
-		deps.Default.Logger, deps.Default.Scope,
 		"{{$spec.EndpointID}}", "{{$spec.HandleID}}", "{{.ThriftService}}::{{.Name}}",
 		{{ if len $middlewares | ne 0 -}}
 			zanzibar.NewTchannelStack([]zanzibar.MiddlewareTchannelHandle{
@@ -94,7 +93,7 @@ func (h *{{$handlerName}}) Handle(
 				zap.String("stacktrace", stacktrace),
 				zap.String("endpoint", h.endpoint.EndpointID))
 
-			h.endpoint.ContextMetrics.IncCounter(ctx, zanzibar.MetricEndpointPanics, 1)
+			h.Deps.Default.ContextMetrics.IncCounter(ctx, zanzibar.MetricEndpointPanics, 1)
 			isSuccessful = false
 			response = nil
 			headers = nil
