+++ runtime/context_test.go
@@ -209,3 +209,63 @@ func TestExtractScope(t *testing.T) {
 	tags := contextExtractor.ExtractScopeTags(ctx)
 	assert.Equal(t, tags, expected)
 }
+func TestLoggerWithFields(t *testing.T) {
+	zapLoggerCore, logs := observer.New(zap.DebugLevel)
+	zapLogger := zap.New(zapLoggerCore)
+	fields := []zap.Field{
+		zap.String("key", "value"),
+	}
+	logger := newLoggerWithFields(zapLogger, fields)
+
+	var logMessages []observer.LoggedEntry
+
+	logger.Debug("msg", zap.String("argField", "argValue"))
+	logMessages = logs.TakeAll()
+	assert.Len(t, logMessages, 1)
+	assert.Equal(t, zap.DebugLevel, logMessages[0].Level)
+	assert.Equal(t, logMessages[0].Context[0].Key, "key")
+	assert.Equal(t, logMessages[0].Context[0].String, "value")
+	assert.Equal(t, logMessages[0].Context[1].Key, "argField")
+	assert.Equal(t, logMessages[0].Context[1].String, "argValue")
+
+	logger.Info("msg", zap.String("argField", "argValue"))
+	logMessages = logs.TakeAll()
+	assert.Len(t, logMessages, 1)
+	assert.Equal(t, zap.InfoLevel, logMessages[0].Level)
+	assert.Equal(t, logMessages[0].Context[0].Key, "key")
+	assert.Equal(t, logMessages[0].Context[0].String, "value")
+	assert.Equal(t, logMessages[0].Context[1].Key, "argField")
+	assert.Equal(t, logMessages[0].Context[1].String, "argValue")
+
+	logger.Warn("msg", zap.String("argField", "argValue"))
+	logMessages = logs.TakeAll()
+	assert.Len(t, logMessages, 1)
+	assert.Equal(t, zap.WarnLevel, logMessages[0].Level)
+	assert.Equal(t, logMessages[0].Context[0].Key, "key")
+	assert.Equal(t, logMessages[0].Context[0].String, "value")
+	assert.Equal(t, logMessages[0].Context[1].Key, "argField")
+	assert.Equal(t, logMessages[0].Context[1].String, "argValue")
+
+	logger.Error("msg", zap.String("argField", "argValue"))
+	logMessages = logs.TakeAll()
+	assert.Len(t, logMessages, 1)
+	assert.Equal(t, zap.ErrorLevel, logMessages[0].Level)
+	assert.Equal(t, logMessages[0].Context[0].Key, "key")
+	assert.Equal(t, logMessages[0].Context[0].String, "value")
+	assert.Equal(t, logMessages[0].Context[1].Key, "argField")
+	assert.Equal(t, logMessages[0].Context[1].String, "argValue")
+
+	assert.Panics(t, func() {
+		logger.Panic("msg", zap.String("argField", "argValue"))
+		logMessages = logs.TakeAll()
+		assert.Len(t, logMessages, 1)
+		assert.Equal(t, zap.ErrorLevel, logMessages[0].Level)
+		assert.Equal(t, logMessages[0].Context[0].Key, "key")
+		assert.Equal(t, logMessages[0].Context[0].String, "value")
+		assert.Equal(t, logMessages[0].Context[1].Key, "argField")
+		assert.Equal(t, logMessages[0].Context[1].String, "argValue")
+	})
+
+	ce := logger.Check(zap.DebugLevel, "msg")
+	assert.NotNil(t, ce)
+}
