+++ runtime/tchannel_client.go
@@ -27,9 +27,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/uber-go/tally"
 	"github.com/uber/tchannel-go"
-	"go.uber.org/thriftrw/protocol"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 	netContext "golang.org/x/net/context"
 )
 
@@ -68,20 +66,6 @@ type TChannelClient struct {
 	routingKey        *string
 	Loggers           map[string]*zap.Logger
 	metrics           ContextMetrics
-}
-
-type tchannelOutboundCall struct {
-	client        *TChannelClient
-	call          *tchannel.OutboundCall
-	methodName    string
-	serviceMethod string
-	success       bool
-	startTime     time.Time
-	finishTime    time.Time
-	reqHeaders    map[string]string
-	resHeaders    map[string]string
-	logger        *zap.Logger
-	metrics       ContextMetrics
 }
 
 // NewTChannelClient is deprecated, use NewTChannelClientContext instead
@@ -274,191 +258,4 @@ func (c *TChannelClient) call(
 	}
 
 	return call.success, call.resHeaders, err
-}
-
-func (c *tchannelOutboundCall) start() {
-	c.startTime = time.Now()
-}
-
-func (c *tchannelOutboundCall) finish(ctx context.Context, err error) {
-	c.finishTime = time.Now()
-
-	// emit metrics
-	if err != nil {
-		errCause := tchannel.GetSystemErrorCode(errors.Cause(err))
-		scopeTags := map[string]string{scopeTagError: errCause.MetricsKey()}
-		ctx = WithScopeTags(ctx, scopeTags)
-		c.metrics.IncCounter(ctx, clientSystemErrors, 1)
-	} else if !c.success {
-		c.metrics.IncCounter(ctx, clientAppErrors, 1)
-	} else {
-		c.metrics.IncCounter(ctx, clientSuccess, 1)
-	}
-	c.metrics.RecordTimer(ctx, clientLatency, c.finishTime.Sub(c.startTime))
-
-	// write logs
-	if err == nil {
-		c.logger.Info("Finished an outgoing client TChannel request", c.logFields()...)
-	} else {
-		c.logger.Warn("Failed to send outgoing client TChannel request", zap.Error(err))
-	}
-}
-
-func (c *tchannelOutboundCall) logFields() []zapcore.Field {
-	var hostPort string
-	if c.call != nil {
-		hostPort = c.call.RemotePeer().HostPort
-	} else {
-		hostPort = "unknown"
-	}
-	fields := []zapcore.Field{
-		zap.String("remoteAddr", hostPort),
-		zap.Time("timestamp-started", c.startTime),
-		zap.Time("timestamp-finished", c.finishTime),
-	}
-
-	for k, v := range c.reqHeaders {
-		fields = append(fields, zap.String("Request-Header-"+k, v))
-	}
-	for k, v := range c.resHeaders {
-		fields = append(fields, zap.String("Response-Header-"+k, v))
-	}
-
-	return fields
-}
-
-// writeReqHeaders writes request headers to arg2
-func (c *tchannelOutboundCall) writeReqHeaders(reqHeaders map[string]string) error {
-	c.reqHeaders = reqHeaders
-
-	twriter, err := c.call.Arg2Writer()
-	if err != nil {
-		return errors.Wrapf(
-			err, "Could not create arg2writer for outbound %s.%s (%s %s) request",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-	if err := WriteHeaders(twriter, reqHeaders); err != nil {
-		_ = twriter.Close()
-		return errors.Wrapf(
-			err, "Could not write headers for outbound %s.%s (%s %s) request",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-	if err := twriter.Close(); err != nil {
-		return errors.Wrapf(
-			err, "Could not close arg2writer for outbound %s.%s (%s %s) request",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-
-	return nil
-}
-
-// writeReqBody writes request body to arg3
-func (c *tchannelOutboundCall) writeReqBody(ctx context.Context, req RWTStruct) error {
-	structWireValue, err := req.ToWire()
-	if err != nil {
-		return errors.Wrapf(
-			err, "Could not write request for outbound %s.%s (%s %s) request",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-
-	twriter, err := c.call.Arg3Writer()
-	if err != nil {
-		return errors.Wrapf(
-			err, "Could not create arg3writer for outbound %s.%s (%s %s) request",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-	if err := protocol.Binary.Encode(structWireValue, twriter); err != nil {
-		_ = twriter.Close()
-		return errors.Wrapf(
-			err, "Could not write request for outbound %s.%s (%s %s) request",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-	if err := twriter.Close(); err != nil {
-		return errors.Wrapf(
-			err, "Could not close arg3writer for outbound %s.%s (%s %s) request",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-
-	// request sent when arg3writer is closed
-	c.metrics.IncCounter(ctx, clientRequest, 1)
-	return nil
-}
-
-// readResHeaders read response headers from arg2
-func (c *tchannelOutboundCall) readResHeaders(response *tchannel.OutboundCallResponse) error {
-	treader, err := response.Arg2Reader()
-	if err != nil {
-		// Do not wrap system errors.
-		if _, ok := err.(tchannel.SystemError); ok {
-			return err
-		}
-		return errors.Wrapf(
-			err, "Could not create arg2reader for outbound %s.%s (%s %s) response",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-	if c.resHeaders, err = ReadHeaders(treader); err != nil {
-		_ = treader.Close()
-		return errors.Wrapf(
-			err, "Could not read headers for outbound %s.%s (%s %s) response",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-	if err := EnsureEmpty(treader, "reading response headers"); err != nil {
-		_ = treader.Close()
-		return errors.Wrapf(
-			err, "Could not ensure arg2reader is empty for outbound %s.%s (%s %s) response",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-	if err := treader.Close(); err != nil {
-		return errors.Wrapf(
-			err, "Could not close arg2reader for outbound %s.%s (%s %s) response",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-
-	// must have called Arg2Reader before calling ApplicationError
-	c.success = !response.ApplicationError()
-	return nil
-}
-
-// readResHeaders read response headers from arg2
-func (c *tchannelOutboundCall) readResBody(response *tchannel.OutboundCallResponse, resp RWTStruct) error {
-	treader, err := response.Arg3Reader()
-	if err != nil {
-		return errors.Wrapf(
-			err, "Could not create arg3Reader for outbound %s.%s (%s %s) response",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-	if err := ReadStruct(treader, resp); err != nil {
-		_ = treader.Close()
-		return errors.Wrapf(
-			err, "Could not read outbound %s.%s (%s %s) response",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-	if err := EnsureEmpty(treader, "reading response body"); err != nil {
-		_ = treader.Close()
-		return errors.Wrapf(
-			err, "Could not ensure arg3reader is empty for outbound %s.%s (%s %s) response",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-	if err := treader.Close(); err != nil {
-		return errors.Wrapf(
-			err, "Could not close arg3reader outbound %s.%s (%s %s) response",
-			c.client.ClientID, c.methodName, c.client.serviceName, c.serviceMethod,
-		)
-	}
-
-	return nil
 }
