+++ test/endpoints/tchannel/baz/baz_metrics_test.go
@@ -195,6 +195,9 @@ func TestCallMetrics(t *testing.T) {
 		"targetendpoint": "SimpleService__call",
 		"dc":             "unknown",
 		"host":           zanzibar.GetHostname(),
+		"device":         "ios",
+		"deviceversion":  "carbon",
+		"regionname":     "san_francisco",
 		"endpointid":     "bazTChannel",
 		"handlerid":      "call",
 		"protocol":       "TChannel",
